refactor: move slack notifier client to provider package
